align_address

Function usize_align_up

pub const fn usize_align_up(addr: usize, align: usize) -> usize

Align address upwards.

Returns the smallest x with alignment align so that x >= addr.

Panics if the alignment is not a power of two or if an overflow occurs.

This is a const version of Align::align_up.