Trait align_address::Align

pub trait Align<A>: Copy {
    // Required methods
    fn align_down(self, align: A) -> Self;
    fn align_up(self, align: A) -> Self;
    fn is_aligned(self, align: A) -> bool;
}

An adress that can be aligned.

Required Methods

fn align_down(self, align: A) -> Self

Align address downwards.

Returns the greatest x with alignment align so that x <= addr.

Panics if the alignment is not a power of two.

fn align_up(self, align: A) -> Self

Align address upwards.

Returns the smallest x with alignment align so that x >= addr.

Panics if the alignment is not a power of two or if an overflow occurs.

fn is_aligned(self, align: A) -> bool

Checks whether the address has the demanded alignment.

Object Safety

This trait is not object safe.