Crate align_address

This crate provides functions for aligning addresses.

Align is implemented for all unsigned integers and provides methods for:

align_down

align_up

is_aligned_to

This crate is based on work from the x86_64 crate, but is available for all architectures and all unsigned integer types.

Examples

use align_address::Align;

assert_eq!(123_u64.align_up(2), 124);

Traits

Align

An adress that can be aligned.

Functions

u8_align_down

Align address downwards.

u8_align_up

Align address upwards.

u8_is_aligned_to

Checks whether the address has the demanded alignment.

u16_align_down

Align address downwards.

u16_align_up

Align address upwards.

u16_is_aligned_to

Checks whether the address has the demanded alignment.

u32_align_down

Align address downwards.

u32_align_up

Align address upwards.

u32_is_aligned_to

Checks whether the address has the demanded alignment.

u64_align_down

Align address downwards.

u64_align_up

Align address upwards.

u64_is_aligned_to

Checks whether the address has the demanded alignment.

u128_align_down

Align address downwards.

u128_align_up

Align address upwards.

u128_is_aligned_to

Checks whether the address has the demanded alignment.

usize_align_down

Align address downwards.

usize_align_up

Align address upwards.

usize_is_aligned_to

Checks whether the address has the demanded alignment.